Fixes clang-tidy issues

+++ tests/unit/Mock.hpp
@@ -10,14 +10,14 @@ namespace Tests::Unit {
 
 struct LoggingSink final : public dlsm::Logger::ISink, std::vector<std::string> {
     std::mutex _records;
-    void AddRecord(std::string&& record) noexcept override {
+    void AddRecord(std::string&& record) noexcept override {  // NOLINT
         std::scoped_lock guard(_records);
         this->emplace_back(record);
     }
     const std::string& tail(std::size_t index = 0) {
         std::scoped_lock guard(_records);
         EXPECT_LT(index, this->size());
-        return *(this->rbegin() + index);
+        return *(this->rbegin() + static_cast<std::ptrdiff_t>(index));
     }
 };
